added role_id & secret_id
